[Write API]: Complete Data classes (InsertClassSectionEvents operation)
