+++ project/src/main/kotlin/org/ionproject/core/writeApi/insertClassSectionEvents/InsertClassSectionEventsRepo.kt
@@ -1,4 +1,4 @@
-package org.ionproject.core.insertClassSectionEvents
+package org.ionproject.core.writeApi.insertClassSectionEvents
 
 import java.sql.Timestamp
